Move date.ts to /common
