+++ extensions/ql-vscode/src/common/date.ts
