+++ extensions/ql-vscode/src/view/common/LastUpdated.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { useMemo } from "react";
 import styled from "styled-components";
 
-import { parseDate } from "../../pure/date";
+import { parseDate } from "../../common/date";
 import { humanizeRelativeTime } from "../../pure/time";
 
 import { Codicon } from "./icon";
