+++ extensions/ql-vscode/src/view/variant-analysis/VariantAnalysisStatusStats.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import styled from "styled-components";
 import { VSCodeLink } from "@vscode/webview-ui-toolkit/react";
-import { formatDate } from "../../pure/date";
+import { formatDate } from "../../common/date";
 import { VariantAnalysisStatus } from "../../variant-analysis/shared/variant-analysis";
 
 export type VariantAnalysisStatusStatsProps = {
