+++ extensions/ql-vscode/test/unit-tests/pure/date.test.ts
@@ -1,4 +1,4 @@
-import { formatDate } from "../../../src/pure/date";
+import { formatDate } from "../../../src/common/date";
 
 describe("Date", () => {
   it("should return a formatted date", () => {
